refactor: added static_cast for vertices to fix build error

+++ FarmGame/src/Graphics/Mesh/Mesh.cpp
@@ -4,7 +4,7 @@
 namespace Engine {
 	Mesh::Mesh(const MeshData& data)
         :  
-        m_VBO(data.vertices.data(), data.vertices.size() * sizeof(Vertex)),
+        m_VBO(data.vertices.data(), static_cast<unsigned int>(data.vertices.size() * sizeof(Vertex))),
 		m_IBO(data.indices.data(), static_cast<unsigned int>(data.indices.size())),
         m_VertexCount(static_cast<uint32_t>(data.vertices.size())),
         m_IndexCount(static_cast<uint32_t>(data.indices.size()))
